Refactor OVS to JIT
