+++ .github/workflows/docker_publish.yml
@@ -54,7 +54,7 @@ jobs:
       - name: Build and push Docker image
         uses: mr-smithers-excellent/docker-build-push@v5
         with:          
-          image: dcsa-ovs-notifications
+          image: dcsa-jit-notifications
           registry: ghcr.io            
           githubOrg: dcsaorg
           username: ${{ github.actor }}
