+++ docker-compose.yml
@@ -1,6 +1,6 @@
 version: '3.5'
 services:
-    dcsa-ovs-notifications:
+    dcsa-jit-notifications:
         build:
            .
         ports:
