+++ pom.xml
@@ -9,10 +9,10 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>org.dcsa</groupId>
-	<artifactId>dcsa-ovs-notifications</artifactId>
+	<artifactId>dcsa-JIT-notifications</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<packaging>jar</packaging>
-	<name>DCSA-OVS-notifications</name>
+	<name>DCSA-JIT-notifications</name>
 	<description>Demo project for Spring Boot</description>
 
 	<properties>
@@ -83,7 +83,7 @@
         <repository>
             <id>github</id>
             <name>GitHub OWNER Apache Maven Packages</name>
-            <url>https://maven.pkg.github.com/dcsaorg/DCSA-OVS</url>
+            <url>https://maven.pkg.github.com/dcsaorg/DCSA-JIT</url>
         </repository>
     </distributionManagement>
