+++ src/main/java/org/dcsa/jit/notifications/Application.java
@@ -1,4 +1,4 @@
-package org.dcsa.ovs.notifications;
+package org.dcsa.jit.notifications;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
