+++ src/main/java/org/dcsa/jit/notifications/model/NotificationEndpoint.java
@@ -1,0 +1,34 @@
+package org.dcsa.jit.notifications.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.relational.core.mapping.Column;
+import org.springframework.data.relational.core.mapping.Table;
+
+import java.util.UUID;
+
+@Table("notification_endpoint")
+@Data
+public class NotificationEndpoint {
+
+    @Id
+    @Column("endpoint_id")
+    private UUID endpointID;
+
+    @Column("subscription_id")
+    private String subscriptionID;
+
+    @Column("secret")
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    private byte[] secret;
+
+    @Column("endpoint_reference")
+    private String endpointReference;
+
+    @Column("managed_endpoint")
+    private Boolean managedEndpoint;
+
+    @Column("subscription_url")
+    private String subscriptionURL;
+}
