+++ src/main/java/org/dcsa/jit/notifications/service/NotificationEndpointService.java
@@ -1,14 +1,11 @@
-package org.dcsa.ovs.notifications.service;
+package org.dcsa.jit.notifications.service;
 
 import org.dcsa.core.service.ExtendedBaseService;
-import org.dcsa.ovs.notifications.model.NotificationEndpoint;
-import org.dcsa.ovs.notifications.util.SubscriberFunction;
+import org.dcsa.jit.notifications.model.NotificationEndpoint;
 import org.springframework.http.server.reactive.ServerHttpRequest;
-import org.springframework.transaction.annotation.Transactional;
 import reactor.core.publisher.Mono;
 
 import java.util.UUID;
-import java.util.function.Consumer;
 
 public interface NotificationEndpointService extends ExtendedBaseService<NotificationEndpoint, UUID> {
