--- src/main/java/org/dcsa/ovs/notifications/model/MailConfiguration.java
@@ -1,35 +1,0 @@
-package org.dcsa.ovs.notifications.model;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.stereotype.Component;
-
-import javax.validation.Valid;
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import java.util.Map;
-
-@Data
-@ConfigurationProperties(prefix = "dcsa.email")
-@Component
-public class MailConfiguration {
-    @Email
-    private String from = "NOT_SPECIFIED";
-    private String timezone = "NOT_SPECIFIED";
-
-    private String dateFormat = "NOT_SPECIFIED";
-
-    private boolean debugEmail = false;
-
-    @Valid
-    private Map<String, MailTemplate> templates;
-
-    public MailTemplate getTemplate(String templateName) {
-        MailTemplate template = templates.get(templateName);
-        if (template == null) {
-            throw new IllegalArgumentException("Unknown template: " + templateName);
-        }
-        return template;
-    }
-}
